renamed WebServer to Websocket for Windows Build
