+++ Sandbox/Sandbox.sln
@@ -7,7 +7,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Server", "Server\Server\Ser
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Client", "Client\Client\Client.vcxproj", "{D7C8DB52-EFFC-4D3C-84A7-B2A4ABEBD899}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WebServer", "WebServer\WebServer\WebServer.vcxproj", "{64C3E63D-BD18-4101-9B58-B46831C58946}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WebSocket", "WebSocket\WebSocket\WebSocket.vcxproj", "{64C3E63D-BD18-4101-9B58-B46831C58946}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
