+++ Sandbox/WebSocket/WebSocket.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30330.147
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WebServer", "WebServer\WebServer.vcxproj", "{32A21EF1-B93E-4DAF-9716-2AC624ACA632}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WebSocket", "WebSocket\WebSocket.vcxproj", "{32A21EF1-B93E-4DAF-9716-2AC624ACA632}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
