+++ Sandbox/WebSocket/WebSocket/WebSocket.vcxproj
@@ -38,7 +38,7 @@
     <VCProjectVersion>16.0</VCProjectVersion>
     <Keyword>Win32Proj</Keyword>
     <ProjectGuid>{64C3E63D-BD18-4101-9B58-B46831C58946}</ProjectGuid>
-    <RootNamespace>WebWebServer</RootNamespace>
+    <RootNamespace>WebWebSocket</RootNamespace>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -376,11 +376,11 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="WebServer\WebServer.cpp" />
+    <ClCompile Include="WebSocket\WebSocket.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="config.h" />
-    <ClInclude Include="WebServer\WebServer.h" />
+    <ClInclude Include="WebSocket\WebSocket.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
