+++ Sandbox/WebSocket/WebSocket/WebSocket.vcxproj.filters
@@ -2,18 +2,18 @@
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="WebServer\WebServer.cpp">
-      <Filter>WebServer</Filter>
+    <ClCompile Include="WebSocket\WebSocket.cpp">
+      <Filter>WebSocket</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="config.h" />
-    <ClInclude Include="WebServer\WebServer.h">
-      <Filter>WebServer</Filter>
+    <ClInclude Include="WebSocket\WebSocket.h">
+      <Filter>WebSocket</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <Filter Include="WebServer">
+    <Filter Include="WebSocket">
       <UniqueIdentifier>{4737ebd7-1139-4821-88d8-908b19177ce4}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
