+++ Win-Net/Net/Net.sln
@@ -15,7 +15,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NetServer", "NetServer\NetS
 		{EFBB82EF-8979-483B-B926-AD76721839AE} = {EFBB82EF-8979-483B-B926-AD76721839AE}
 	EndProjectSection
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NetWebServer", "NetWebServer\NetWebServer.vcxproj", "{80F9040F-0B79-49E9-9628-54D1C9327123}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NetWebSocket", "NetWebSocket\NetWebSocket.vcxproj", "{80F9040F-0B79-49E9-9628-54D1C9327123}"
 	ProjectSection(ProjectDependencies) = postProject
 		{EFBB82EF-8979-483B-B926-AD76721839AE} = {EFBB82EF-8979-483B-B926-AD76721839AE}
 	EndProjectSection
