+++ Win-Net/Net/NetWebSocket/AssemblyInfo.cpp
