+++ Win-Net/Net/NetWebSocket/NetWebSocket.vcxproj
@@ -57,6 +57,7 @@
     <Keyword>ManagedCProj</Keyword>
     <RootNamespace>NetWebServer</RootNamespace>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
+    <ProjectName>NetWebSocket</ProjectName>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='DEBUG-STATIC|Win32'" Label="Configuration">
